Refactor: DailyMealUtils updated
